+++ src/components/app/layout/questions/carousel/question-carousel-card.tsx
