+++ src/components/app/layout/questions/carousel/question-carousel.tsx
